requirements.txt
flask==3.0.2
selenium==4.18.1
python-dotenv==1.0.1
gunicorn==21.2.0
webdriver-manager==4.0.1
pandas==2.2.1
nltk==3.8.1
scikit-learn==1.3.2
beautifulsoup4==4.12.3
requests==2.31.0 